ocr.pytorch/requirements.txt
#aniso8601==9.0.1
#certifi==2022.5.18.1
#charset-normalizer
#click==8.1.3
Flask
Flask-RESTful
#gunicorn
#idna
#importlib-metadata
#itsdangerous
Jinja2
#MarkupSafe
nltk
numpy
opencv-python
pandas
Pillow
#preprocessing
python-dateutil
#pytz
requests
six
#sphinx-rtd-theme
torch #==1.10.1
torchvision
#typing_extensions
urllib3
Werkzeug
zipp
